requirements.txt
# Stats Spark - Production Dependencies
# GitHub Profile Statistics Generator

# GitHub API client
PyGithub>=2.1.1

# Configuration parsing
PyYAML>=6.0.1

# SVG generation
svgwrite>=1.4.3

# HTTP requests with retry logic
requests>=2.31.0

# Date and time utilities
python-dateutil>=2.8.2
